Revert "Configuração de Migration e implementação de views e controllers de todos os models"
This reverts commit df3b11639e4f19f5bbac6fa51c9172a84c95e99a.

+++ Tecidolandia/Models/Entities/Cliente.cs
@@ -13,7 +13,7 @@ namespace Tecidolandia.Models.Entities
         [Key]
         [Required(ErrorMessage = "O campo nome é obrigatório")]
         [Column("ID_CLIENTE")]
-        public Int64 IdCliente { get; set; }
+        public int IdCliente { get; set; }
 
         [Column("NM_COMPLETO")]
         [Display(Name = "Nome Completo")]
